MuseIA/requirements.txt
streamlit
requests
pandas
beautifulsoup4
pdfplumber
google-generativeai==0.7.2
supabase
bcrypt
huggingface
python-dotenv
python-dotenv
mercadopago